Promote ov002/daChRoom_c to the single TU src/actors/d_a_ch_room.cpp
Eight per-function objects become one. The cartridge's contiguous linker run
.text 0x020b07f8..0x020b09b0 is now a genuine translation unit -- the
production build links this object rather than eight -- and the eight legacy
delinks entries collapse into the single `complete` span. The filename is
tools/tu_names.py's: candidate_stem('daChRoom_c') is d_a_ch_room.

WHAT MAKES THE PROMOTION LEGAL IS THE INLINE DESTRUCTOR, and what makes the
inline destructor legal is a measurement. Declared out of line, mwccarm
2004/b56 emits D0 before D1 -- the reverse of the cartridge, which has D1 at
0x020b07f8 and D0 at 0x020b081c -- and adds a third, homeless D2 that no ROM
address claims; tools/objisolate.py then refuses the whole TU. Defined inline
in the header, the pair comes out in ROM order and there is no D2. That is only
safe for a leaf, and this class is one by measurement rather than by assumption:
the word 0x021085b8 (_ZTI10daChRoom_c) occurs exactly once anywhere under
extracted/ -- ov002 file offset 0x5af94, which is 0x021085f4, the typeinfo slot
of its own vtable header at _ZTV-4 (the symbols.txt address 0x021085f8 is the
address point). The two hits the scan reports are the same 394048-byte image
under two paths. No other class's __si_class_type_info points at it, so no
descendant exists to inline this body where the ROM has a `bl`.

THE BODY REALLY IS EMPTY. D1 is 0x24 bytes and disassembles to exactly
`push {r4,lr} / ldr r1,[pc,#0x14] / mov r4,r0 / str r1,[r4] / bl 0x020112c8 /
mov r0,r4 / pop {r4,lr} / bx lr`: one vptr store of its own, then the tail into
~dActor_c. There is no member destructor call to account for, and the header
explains why -- this class declares no member at all, every field it uses is an
inherited dActor_c one. Contrast daObjWakame_c, whose inline `{}` compiles to
0x30 because it owns a ModelAnim; an inline empty destructor is not the same
thing as an empty destructor.

THE OVERRIDES ARE THE VTABLE DIFF, not a guess. _ZTV10daChRoom_c and
_ZTV8dActor_c (arm9 0x0208e3a4) are 31 words each and differ in exactly slots
0, 3, 6, 9, 12, 16 and 17 -- InitResources, CleanupResources, Behavior, Render,
OnPendingDestroy, D1 and D0, precisely the seven members the header declares.
Every other slot holds the base's own word and is deliberately not redeclared.

NAMING THE CLASS PUTS ITS VAGUE-LINKAGE DATA IN THIS OBJECT, because this TU
defines the key function. Nine records come out; tools/romdata_check.py compares
each against the cartridge with relocations applied, before production isolation
discards it:

  _ZTV10daChRoom_c    ov002 0x021085f8  VERIFIED, 124 bytes = 31 slots
  _ZTI10daChRoom_c    ov002 0x021085b8  VERIFIED
  _ZTI8dActor_c / _ZTI7dBase_c / _ZTI7fBase_c   arm9   VERIFIED
  the four _ZTS records                                PARTIAL
  nothing                                              DIFFERS

The four PARTIAL rows are the known dsd extent shortfall on _ZTS records -- a
range that stops short, not a disagreement about bytes. A tenth symbol,
_ZN7Vector3D1Ev, is licensed as deadstrip-duplicate rather than deadstrip-data:
it is a function with a real ROM home at arm9 0x020072c0 that this object
re-emits, the same disposition daObjCannonShutter_c uses.

106/106 exact, 100.000000%, mismatching 0. tools/tiers_ratchet.py --update
reports 2592 -> 2592, +0/-0: the six CONVERTED members move from their legacy
per-function paths to this file, a rename within the set rather than a
backslide. layout-check clean; romdata_check's 5 DIFFERS rows are pre-existing
and belong to other classes.
