+++ config/converted-baseline.json
@@ -507,12 +507,6 @@
     "src/_ZN11SnowmanHead8SetStateEi.cpp",
     "src/_ZN11SnowmanHeadD0Ev.cpp",
     "src/_ZN11SnowmanHeadD1Ev.cpp",
-    "src/_ZN10daChRoom_c13InitResourcesEv.cpp",
-    "src/_ZN10daChRoom_c16CleanupResourcesEv.cpp",
-    "src/_ZN10daChRoom_c16OnPendingDestroyEv.cpp",
-    "src/_ZN10daChRoom_c6RenderEv.cpp",
-    "src/_ZN10daChRoom_cD0Ev.cpp",
-    "src/_ZN10daChRoom_cD1Ev.cpp",
     "src/_ZN11VolcanoFire11ChangeStateEPNS_5StateE.cpp",
     "src/_ZN11VolcanoFireD0Ev.cpp",
     "src/_ZN11VolcanoFireD1Ev.cpp",
@@ -2415,6 +2409,12 @@
     "src/actors/d_a_cam_tag.cpp#_ZN10daCamTag_c8BehaviorEv",
     "src/actors/d_a_cam_tag.cpp#_ZN10daCamTag_cD0Ev",
     "src/actors/d_a_cam_tag.cpp#_ZN10daCamTag_cD1Ev",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_c13InitResourcesEv",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_c16CleanupResourcesEv",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_c16OnPendingDestroyEv",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_c6RenderEv",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_cD0Ev",
+    "src/actors/d_a_ch_room.cpp#_ZN10daChRoom_cD1Ev",
     "src/actors/d_a_i_donketu.cpp#_ZN12daIDonketu_c14UpdateRunStateEv",
     "src/actors/d_a_i_donketu.cpp#_ZN12daIDonketu_c16UpdateDeathStateEv",
     "src/actors/d_a_i_donketu.cpp#_ZN12daIDonketu_c8BehaviorEv",
