--- src/_ZN10daChRoom_c16CleanupResourcesEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN10daChRoom_c16CleanupResourcesEv
-
-#include "daChRoom_c.h"
-
-int daChRoom_c::CleanupResources()
-{
-    return 1;
-}
