--- src/_ZN10daChRoom_c16OnPendingDestroyEv.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN10daChRoom_c16OnPendingDestroyEv
-
-#include "daChRoom_c.h"
-
-void daChRoom_c::OnPendingDestroy()
-{
-}
