--- src/_ZN10daChRoom_c6RenderEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN10daChRoom_c6RenderEv
-
-#include "daChRoom_c.h"
-
-int daChRoom_c::Render()
-{
-    return 1;
-}
