--- src/_ZN10daChRoom_cD1Ev.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN10daChRoom_cD1Ev
-
-#include "daChRoom_c.h"
-
-daChRoom_c::~daChRoom_c()
-{
-}
