--- src_tu/actors/daChRoom_c.cpp
@@ -1,155 +1,0 @@
-//cpp
-/* SHADOW translation unit -- generated by tools/tubuild.py create.
- * ov002/daChRoom_c  (8 function(s))
- *
- * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
- * build. It is a STARTING POINT (plan sec 7.3): local shadow declarations
- * below were carried verbatim from the legacy files, not reconciled against
- * real project headers -- that judgement call is left to a human/LLM review,
- * the way pilot #1 reconciled PoleLift::Render and ::CleanupResources by hand
- * (see notes/tu-reconstruction-pilot-report.md sec 5.2).
- *
- * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
- * emits one .text section per function, in the REVERSE of source order, so
- * the highest-address ROM function is written FIRST here. Do not reorder;
- * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
- * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
- *
- * Assembled from these legacy one-function sources (ROM address order):
- *   [0] 0x020b07f8  src/_ZN10daChRoom_cD1Ev.cpp
- *   [1] 0x020b081c  src/_ZN10daChRoom_cD0Ev.cpp
- *   [2] 0x020b0854  src/_ZN10daChRoom_c16CleanupResourcesEv.cpp
- *   [3] 0x020b085c  src/_ZN10daChRoom_c16OnPendingDestroyEv.cpp
- *   [4] 0x020b0860  src/_ZN10daChRoom_c6RenderEv.cpp
- *   [5] 0x020b0868  src/_ZN10daChRoom_c8BehaviorEv.cpp
- *   [6] 0x020b0938  src/_ZN10daChRoom_c13InitResourcesEv.cpp
- *   [7] 0x020b0980  src/daChRoom_c_Spawn.c
- */
-
-/* Includes: union of the legacy files', first-seen in ROM-ascending
- * processing order. NOT verified for header ordering constraints (e.g. a
- * common.h-before-X rule) -- watch for new compile errors after this. */
-#include "daChRoom_c.h"
-#include "types.h"
-#include "decl_common.h"
-#include "decl_Actor.h"
-#include "decl_ActorBase.h"
-
-/* Local shadow declarations carried from the legacy files verbatim.
- * NOT reconciled against real project headers -- check include/*.h for
- * each of these before compiling; a real header should usually win. */
-extern "C" {
-extern void Vec3_Sub(Vector3* out, Vector3* a, Vector3* b);
-extern void Vec3_RotateYAndTranslate(Vector3* out, void* m, s16 ang, Vector3* in);
-extern u8 data_0209f250;
-extern char* data_0209f394[];
-extern char data_020a0ebc;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 7 -- daChRoom_c_Spawn, 0x020b0980, size 0x30 */
-/* -------------------------------------------------------------------------- */
-// @symbol daChRoom_c_Spawn
-/* recovered: globals resolved, declarations from a shared header */
-/* recovered: globals resolved */
-/* resolved: VT = _ZTV10daChRoom_c */
-extern "C" {  /* .c-derived member: C linkage for the whole block */
-int *daChRoom_c_Spawn(void)
-{
-    int *p = (int *)_ZN7fBase_cnwEj(sizeof(daChRoom_c));
-    if (p) { _ZN8dActor_cC2Ev(p); p[0] = (int)(_ZTV10daChRoom_c + 2); }
-    return p;
-}
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 6 -- _ZN10daChRoom_c13InitResourcesEv, 0x020b0938, size 0x48 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_c13InitResourcesEv
-/* recovered: named members + shared header, real C++ method */
-int daChRoom_c::InitResources()
-{
-  mScaleX=(((param1&0xf)+1)*0x64000)>>1;
-  mScaleY=(((param1>>4&0xf)+1)*0x64000);
-  mAngleY=-mAngleY;
-  return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 5 -- _ZN10daChRoom_c8BehaviorEv, 0x020b0868, size 0xd0 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_c8BehaviorEv
-/* recovered: named members + shared header, real C++ method, declarations from a shared header */
-/* recovered: named members + shared header, real C++ method */
-int daChRoom_c::Behavior()
-{
-    char* obj;
-    Vector3 d;
-    Vector3 r;
-    int v;
-
-    obj = data_0209f394[data_0209f250];
-    Vec3_Sub(&d, (Vector3*)(obj + 0x5c), (Vector3*)&mPosX);
-    Vec3_RotateYAndTranslate(&r, &data_020a0ebc, mAngleY, &d);
-
-    v = r.x;
-    if (v < 0) v = -v;
-    if (v < mScaleX) {
-        if (r.y > -0x96000) {
-            if (r.y < mScaleY) {
-                int z = r.z;
-                int az = (z < 0) ? -z : z;
-                if (az > 0x64000 && az < 0x190000) {
-                    int area = (z < 0) ? mAngleX : mAngleZ;
-                    *(char*)(obj + 0xcc) = (char)area;
-                    ChangeArea((char)area);
-                }
-            }
-        }
-    }
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 4 -- _ZN10daChRoom_c6RenderEv, 0x020b0860, size 0x8 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_c6RenderEv
-int daChRoom_c::Render()
-{
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 3 -- _ZN10daChRoom_c16OnPendingDestroyEv, 0x020b085c, size 0x4 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_c16OnPendingDestroyEv
-void daChRoom_c::OnPendingDestroy()
-{
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 2 -- _ZN10daChRoom_c16CleanupResourcesEv, 0x020b0854, size 0x8 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_c16CleanupResourcesEv
-int daChRoom_c::CleanupResources()
-{
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 1 -- _ZN10daChRoom_cD0Ev, 0x020b081c, size 0x38 */
-/* -------------------------------------------------------------------------- */
-/* NOT hand-written: the legacy D0 and D1 files each independently
- * reconstructed the same `daChRoom_c::~daChRoom_c(){}` as a real
- * out-of-line destructor -- the standard D0/D1 collapse artifact of one
- * destructor split across two one-function files. Keeping both here
- * redefines the same symbol; one definition below (D1) is enough, and the
- * compiler emits both D0 and D1 from it. */
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 0 -- _ZN10daChRoom_cD1Ev, 0x020b07f8, size 0x24 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN10daChRoom_cD1Ev
-daChRoom_c::~daChRoom_c()
-{
-}
